chore: disable annoying ts/eslint rules
ts/no-misused-promises was complaining with react-hook-form when I did something like ` <form onSubmit={handleSubmit(onSubmit)}>`. Since handleSubmit returns a promise, it was in violation of the rule. Turning off checks for when the called function returns void addresses the annoyance.

ts/consistent-type-definitions was disabled because I like the flexibility of using both types and interfaces across a codebase.